+++ packages/enterprise/src/license.ts
@@ -40,6 +40,7 @@ export type CommercialFeature =
   | "no-access-role"
   | "teams"
   | "sticky-bucketing"
+  | "require-approvals"
   | "code-references"
   | "prerequisites"
   | "prerequisite-targeting"
@@ -52,6 +53,7 @@ export interface LicenseInterface {
   companyName: string; // Name of the organization on the license
   organizationId?: string; // OrganizationId (keys prior to 12/2022 do not contain this field)
   seats: number; // Maximum number of seats on the license
+  hardCap: boolean; // True if this license has a hard cap on the number of seats
   dateCreated: string; // Date the license was issued
   dateExpires: string; // Date the license expires
   isTrial: boolean; // True if this is a trial license
@@ -72,7 +74,7 @@ export interface LicenseInterface {
   signedChecksum: string; // Checksum of the license data signed with the private key
 }
 
-// Old style license keys where the license data is encrypted in the key itself
+// Old/Airgapped style license keys where the license data is encrypted in the key itself
 type LicenseData = {
   // Unique id for the license key
   ref: string;
@@ -82,6 +84,8 @@ type LicenseData = {
   org?: string;
   // Max number of seats
   qty: number;
+  // True if this license has a hard cap on the number of seats (keys prior to 03/2024 do not contain this field)
+  hardCap?: boolean;
   // Date issued
   iat: string;
   // Expiration date
@@ -163,6 +167,7 @@ export const accountFeatures: CommercialFeaturesMap = {
     "custom-launch-checklist",
     "no-access-role",
     "sticky-bucketing",
+    "require-approvals",
     "code-references",
     "prerequisites",
     "prerequisite-targeting",
@@ -278,6 +283,7 @@ export async function getVerifiedLicenseData(
     companyName: decodedLicense.sub,
     organizationId: decodedLicense.org,
     seats: decodedLicense.qty,
+    hardCap: decodedLicense.hardCap || false,
     dateCreated: decodedLicense.iat,
     dateExpires: decodedLicense.exp,
     isTrial: decodedLicense.trial,