+++ packages/front-end/pages/report/[rid].tsx
@@ -16,9 +16,6 @@ import RunQueriesButton, {
   getQueryStatus,
 } from "@/components/Queries/RunQueriesButton";
 import DateResults from "@/components/Experiment/DateResults";
-import BreakDownResults_old from "@/components/Experiment/BreakDownResults_old";
-import CompactResults_old from "@/components/Experiment/CompactResults_old";
-import GuardrailResults from "@/components/Experiment/GuardrailResult";
 import { useAuth } from "@/services/auth";
 import ControlledTabs from "@/components/Tabs/ControlledTabs";
 import Tab from "@/components/Tabs/Tab";
@@ -38,26 +35,20 @@ import Tooltip from "@/components/Tooltip/Tooltip";
 import { useUser } from "@/services/UserContext";
 import VariationIdWarning from "@/components/Experiment/VariationIdWarning";
 import DeleteButton from "@/components/DeleteButton/DeleteButton";
-import PValueGuardrailResults from "@/components/Experiment/PValueGuardrailResults";
 import useOrgSettings from "@/hooks/useOrgSettings";
-import track, { trackReport } from "@/services/track";
+import { trackReport } from "@/services/track";
 import CompactResults from "@/components/Experiment/CompactResults";
-import { useLocalStorage } from "@/hooks/useLocalStorage";
 import BreakDownResults from "@/components/Experiment/BreakDownResults";
 import DimensionChooser from "@/components/Dimensions/DimensionChooser";
+import PageHead from "@/components/Layout/PageHead";
 
 export default function ReportPage() {
-  const [newUi, setNewUi] = useLocalStorage<boolean>(
-    "experiment-results-new-ui-v2",
-    true
-  );
-
   const router = useRouter();
   const { rid } = router.query;
 
   const [editModalOpen, setEditModalOpen] = useState(false);
 
-  const { getExperimentMetricById, getDatasourceById } = useDefinitions();
+  const { getDatasourceById } = useDefinitions();
   const { data, error, mutate } = useApi<{ report: ReportInterface }>(
     `/report/${rid}`
   );
@@ -150,26 +141,21 @@ export default function ReportPage() {
 
   return (
     <>
-      <div
-        className="alert-secondary p-2 mb-2 text-center"
-        style={{ marginTop: -5 }}
-      >
-        You are using the {newUi ? "new" : "old"} experiment results view.{" "}
-        <a
-          role="button"
-          className="a"
-          onClick={() => {
-            track("Switched Experiment Results UI", {
-              switchTo: newUi ? "old" : "new",
-            });
-            setNewUi(!newUi);
-          }}
-        >
-          {newUi
-            ? "Switch back to the old view?"
-            : "Try the new experiment results view?"}
-        </a>
-      </div>
+      <PageHead
+        breadcrumb={[
+          {
+            display: `Experiments`,
+            href: `/experiments`,
+          },
+          {
+            display: `${experimentData?.experiment.name ?? "Report"}`,
+            href: experimentData?.experiment.id
+              ? `/experiment/${experimentData.experiment.id}`
+              : undefined,
+          },
+          { display: report.title },
+        ]}
+      />
       <div className="container-fluid pagecontents experiment-details">
         {editModalOpen && (
           <Modal
@@ -217,7 +203,7 @@ export default function ReportPage() {
         <div className="mb-3">
           {report?.experimentId && (
             <Link href={`/experiment/${report.experimentId}#results`}>
-              <GBCircleArrowLeft />
+              <GBCircleArrowLeft className="mr-2" />
               Go to experiment results
             </Link>
           )}
@@ -296,9 +282,21 @@ export default function ReportPage() {
                     exposureQueryId={report.args.exposureQueryId}
                     userIdType={report.args.userIdType}
                     labelClassName="mr-2"
-                    newUi={true}
                     disabled={true}
                   />
+                </div>
+                <div className="col-auto d-flex align-items-end mr-3">
+                  <div>
+                    <div className="uppercase-title text-muted">Date range</div>
+                    <div className="relative">
+                      <span className="date-label">
+                        {date(report.args.startDate)} —{" "}
+                        {report.args.endDate
+                          ? date(report.args.endDate)
+                          : "now"}
+                      </span>
+                    </div>
+                  </div>
                 </div>
                 <div className="col-auto">
                   {hasData &&
@@ -386,12 +384,6 @@ export default function ReportPage() {
                     }}
                     supportsNotebooks={!!datasource?.settings?.notebookRunQuery}
                     // @ts-expect-error TS(2322) If you come across this, please fix it!: Type '(() => void) | null' is not assignable to ty... Remove this comment to see the full error message
-                    configure={
-                      permissions.check("createAnalyses", "")
-                        ? () => setActive("Configuration")
-                        : null
-                    }
-                    // @ts-expect-error TS(2322) If you come across this, please fix it!: Type '(() => void) | null' is not assignable to ty... Remove this comment to see the full error message
                     editMetrics={
                       permissions.check("createAnalyses", "")
                         ? () => setActive("Configuration")
@@ -461,7 +453,7 @@ export default function ReportPage() {
                   variations={variations}
                   statsEngine={report.args.statsEngine}
                 />
-              ) : newUi ? (
+              ) : (
                 <BreakDownResults
                   isLatestPhase={true}
                   metrics={report.args.metrics}
@@ -484,29 +476,6 @@ export default function ReportPage() {
                   }
                   sequentialTestingEnabled={sequentialTestingEnabled}
                   differenceType={"relative"}
-                />
-              ) : (
-                <BreakDownResults_old
-                  isLatestPhase={true}
-                  metrics={report.args.metrics}
-                  // @ts-expect-error TS(2322) If you come across this, please fix it!: Type 'MetricOverride[] | undefined' is not assigna... Remove this comment to see the full error message
-                  metricOverrides={report.args.metricOverrides}
-                  reportDate={report.dateCreated}
-                  results={report.results?.dimensions || []}
-                  status={"stopped"}
-                  startDate={getValidDate(report.args.startDate).toISOString()}
-                  dimensionId={report.args.dimension}
-                  activationMetric={report.args.activationMetric}
-                  guardrails={report.args.guardrails}
-                  variations={variations}
-                  key={report.args.dimension}
-                  statsEngine={report.args.statsEngine}
-                  pValueCorrection={pValueCorrection}
-                  regressionAdjustmentEnabled={regressionAdjustmentEnabled}
-                  metricRegressionAdjustmentStatuses={
-                    report.args.metricRegressionAdjustmentStatuses
-                  }
-                  sequentialTestingEnabled={sequentialTestingEnabled}
                 />
               ))}
             {report.results && !report.args.dimension && (
@@ -548,110 +517,35 @@ export default function ReportPage() {
             {hasData &&
               !report.args.dimension &&
               report.results?.dimensions?.[0] !== undefined && (
-                <>
-                  {newUi ? (
-                    <div className="mt-0 mb-3">
-                      <CompactResults
-                        variations={variations}
-                        multipleExposures={
-                          report.results?.multipleExposures || 0
-                        }
-                        results={report.results?.dimensions?.[0]}
-                        queryStatusData={queryStatusData}
-                        reportDate={report.dateCreated}
-                        startDate={getValidDate(
-                          report.args.startDate
-                        ).toISOString()}
-                        isLatestPhase={true}
-                        status={"stopped"}
-                        metrics={report.args.metrics}
-                        metricOverrides={report.args.metricOverrides ?? []}
-                        guardrails={report.args.guardrails}
-                        id={report.id}
-                        statsEngine={
-                          report.args.statsEngine || DEFAULT_STATS_ENGINE
-                        }
-                        pValueCorrection={pValueCorrection}
-                        regressionAdjustmentEnabled={
-                          regressionAdjustmentEnabled
-                        }
-                        metricRegressionAdjustmentStatuses={
-                          report.args.metricRegressionAdjustmentStatuses
-                        }
-                        sequentialTestingEnabled={sequentialTestingEnabled}
-                        differenceType={"relative"}
-                        isTabActive={true}
-                      />
-                    </div>
-                  ) : (
-                    <>
-                      <CompactResults_old
-                        id={report.id}
-                        isLatestPhase={true}
-                        metrics={report.args.metrics}
-                        // @ts-expect-error TS(2322) If you come across this, please fix it!: Type 'MetricOverride[] | undefined' is not assigna... Remove this comment to see the full error message
-                        metricOverrides={report.args.metricOverrides}
-                        reportDate={report.dateCreated}
-                        results={report.results?.dimensions?.[0]}
-                        status={"stopped"}
-                        startDate={getValidDate(
-                          report.args.startDate
-                        ).toISOString()}
-                        multipleExposures={
-                          report.results?.multipleExposures || 0
-                        }
-                        variations={variations}
-                        statsEngine={report.args.statsEngine}
-                        pValueCorrection={pValueCorrection}
-                        regressionAdjustmentEnabled={
-                          regressionAdjustmentEnabled
-                        }
-                        metricRegressionAdjustmentStatuses={
-                          report.args.metricRegressionAdjustmentStatuses
-                        }
-                        sequentialTestingEnabled={sequentialTestingEnabled}
-                      />
-                      {/* @ts-expect-error TS(2532) If you come across this, please fix it!: Object is possibly 'undefined'. */}
-                      {report.args.guardrails?.length > 0 && (
-                        <div className="mt-1 px-3">
-                          <h3 className="mb-3">Guardrails</h3>
-                          <div className="row">
-                            {/* @ts-expect-error TS(2532) If you come across this, please fix it!: Object is possibly 'undefined'. */}
-                            {report.args.guardrails.map((g) => {
-                              const metric = getExperimentMetricById(g);
-                              if (!metric) return "";
-
-                              const data =
-                                report.results?.dimensions?.[0]?.variations;
-                              if (!data) return "";
-
-                              return (
-                                <div
-                                  className="col-12 col-xl-4 col-lg-6 mb-3"
-                                  key={g}
-                                >
-                                  {report.args.statsEngine === "frequentist" ? (
-                                    <PValueGuardrailResults
-                                      data={data}
-                                      variations={variations}
-                                      metric={metric}
-                                    />
-                                  ) : (
-                                    <GuardrailResults
-                                      data={data}
-                                      variations={variations}
-                                      metric={metric}
-                                    />
-                                  )}
-                                </div>
-                              );
-                            })}
-                          </div>
-                        </div>
-                      )}
-                    </>
-                  )}
-                </>
+                <div className="mt-0 mb-3">
+                  <CompactResults
+                    variations={variations}
+                    multipleExposures={report.results?.multipleExposures || 0}
+                    results={report.results?.dimensions?.[0]}
+                    queryStatusData={queryStatusData}
+                    reportDate={report.dateCreated}
+                    startDate={getValidDate(
+                      report.args.startDate
+                    ).toISOString()}
+                    isLatestPhase={true}
+                    status={"stopped"}
+                    metrics={report.args.metrics}
+                    metricOverrides={report.args.metricOverrides ?? []}
+                    guardrails={report.args.guardrails}
+                    id={report.id}
+                    statsEngine={
+                      report.args.statsEngine || DEFAULT_STATS_ENGINE
+                    }
+                    pValueCorrection={pValueCorrection}
+                    regressionAdjustmentEnabled={regressionAdjustmentEnabled}
+                    metricRegressionAdjustmentStatuses={
+                      report.args.metricRegressionAdjustmentStatuses
+                    }
+                    sequentialTestingEnabled={sequentialTestingEnabled}
+                    differenceType={"relative"}
+                    isTabActive={true}
+                  />
+                </div>
               )}
             {hasData && (
               <div className="row align-items-center mx-2 my-3">