+++ packages/sdk-js/src/GrowthBook.ts
@@ -15,6 +15,7 @@ import type {
   LoadFeaturesOptions,
   RealtimeUsageData,
   RefreshFeaturesOptions,
+  RenderFunction,
   Result,
   StickyAssignments,
   StickyAssignmentsDocument,
@@ -114,6 +115,10 @@ export class GrowthBook<
     this._loadFeaturesCalled = false;
     this._redirectedUrl = "";
     this._deferredTrackingCalls = [];
+
+    if (context.renderer) {
+      this._renderer = context.renderer;
+    }
 
     if (context.remoteEval) {
       if (context.decryptionKey) {
@@ -233,7 +238,11 @@ export class GrowthBook<
 
   private _render() {
     if (this._renderer) {
-      this._renderer();
+      try {
+        this._renderer();
+      } catch (e) {
+        console.error("Failed to render", e);
+      }
     }
   }
 
@@ -316,6 +325,10 @@ export class GrowthBook<
     }
     this._render();
     this._updateAllAutoExperiments();
+  }
+
+  public async updateAttributes(attributes: Attributes) {
+    return this.setAttributes({ ...this._ctx.attributes, ...attributes });
   }
 
   public async setAttributeOverrides(overrides: Attributes) {
@@ -434,7 +447,7 @@ export class GrowthBook<
     this._triggeredExpKeys.clear();
   }
 
-  public setRenderer(renderer: () => void) {
+  public setRenderer(renderer: null | RenderFunction) {
     this._renderer = renderer;
   }