+++ packages/sdk-js/src/auto-wrapper.ts
@@ -3,12 +3,17 @@ import { Context, GrowthBook } from "./index";
 declare global {
   interface Window {
     _growthbook?: GrowthBook;
+    growthbook_queue?:
+      | Array<(gb: GrowthBook) => void>
+      | { push: (cb: (gb: GrowthBook) => void) => void };
     growthbook_config?: Context;
     // eslint-disable-next-line
-    dataLayer: any[];
+    dataLayer?: any[];
     analytics?: {
       track?: (name: string, props?: Record<string, unknown>) => void;
     };
+    // eslint-disable-next-line
+    gtag?: (...args: any) => void;
   }
 }
 
@@ -83,6 +88,32 @@ function getUtmAttributes() {
   return utms;
 }
 
+function getDataLayerVariables() {
+  if (!window.dataLayer || !window.dataLayer.forEach) return {};
+  const obj: Record<string, unknown> = {};
+  window.dataLayer.forEach((item: unknown) => {
+    // Skip empty and non-object entries
+    if (!item || typeof item !== "object" || "length" in item) return;
+
+    // Skip events
+    if ("event" in item) return;
+
+    Object.keys(item).forEach((k) => {
+      // Filter out known properties that aren't useful
+      if (typeof k !== "string" || k.match(/^(gtm)/)) return;
+
+      const val = (item as Record<string, unknown>)[k];
+
+      // Only add primitive variable values
+      const valueType = typeof val;
+      if (["string", "number", "boolean"].includes(valueType)) {
+        obj[k] = val;
+      }
+    });
+  });
+  return obj;
+}
+
 function getAutoAttributes() {
   const ua = navigator.userAgent;
 
@@ -97,19 +128,18 @@ function getAutoAttributes() {
     : "unknown";
 
   return {
+    ...getDataLayerVariables(),
     id: getUUID(),
     url: location.href,
     path: location.pathname,
     host: location.host,
     query: location.search,
+    pageTitle: document && document.title,
     deviceType: ua.match(/Mobi/) ? "mobile" : "desktop",
     browser,
     ...getUtmAttributes(),
   };
 }
-
-// Initialize the data layer if it doesn't exist yet (GA4, GTM)
-window.dataLayer = window.dataLayer || [];
 
 const currentScript = document.currentScript;
 const dataContext = currentScript ? currentScript.dataset : {};
@@ -131,7 +161,14 @@ const gb = new GrowthBook({
   subscribeToChanges: true,
   trackingCallback: (e, r) => {
     const p = { experiment_id: e.key, variation_id: r.key };
-    window.dataLayer.push(["event", "experiment_viewed", p]);
+
+    // GA4 (gtag and GTM options)
+    window.gtag
+      ? window.gtag("event", "experiment_viewed", p)
+      : window.dataLayer &&
+        window.dataLayer.push({ event: "experiment_viewed", ...p });
+
+    // Segment
     window.analytics &&
       window.analytics.track &&
       window.analytics.track("Experiment Viewed", p);
@@ -149,12 +186,41 @@ setInterval(() => {
   if (location.href !== currentUrl) {
     currentUrl = location.href;
     gb.setURL(currentUrl);
-    gb.setAttributes({
-      ...gb.getAttributes(),
-      ...getAttributes(),
-    });
+    gb.updateAttributes(getAttributes());
   }
 }, 500);
+
+// Listen for a custom event to update URL and attributes
+document.addEventListener("growthbookrefresh", () => {
+  if (location.href !== currentUrl) {
+    currentUrl = location.href;
+    gb.setURL(currentUrl);
+  }
+  gb.updateAttributes(getAttributes());
+});
+
+const fireCallback = (cb: (gb: GrowthBook) => void) => {
+  try {
+    cb && cb(gb);
+  } catch (e) {
+    console.error("Uncaught growthbook_queue error", e);
+  }
+};
+
+// Process any queued callbacks
+if (window.growthbook_queue) {
+  if (Array.isArray(window.growthbook_queue)) {
+    window.growthbook_queue.forEach((cb) => {
+      fireCallback(cb);
+    });
+  }
+}
+// Replace the queue with a function that immediately calls the callback
+window.growthbook_queue = {
+  push: (cb: (gb: GrowthBook) => void) => {
+    fireCallback(cb);
+  },
+};
 
 // Store a reference in window to enable more advanced use cases
 export default gb;